refactor: moving the `validate.LowerCaseString` into the `privatedns` package since this is the only place it's used
